requirements.txt
PyQt6
websocket-client
requests
numpy
scipy
langchain
langchain-openai
langchain-community
langchain-core
python-docx
pypdf
scikit-learn
pandas
openpyxl
xlrd
